refactor: move copy address button to own component
